requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
coverage==5.0.2
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-nested-admin==3.2.4
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
httpie==1.0.3
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
gunicorn==19.7.1
whitenoise==3.3.1
